create directory for vectorize.pkl file

+++ src/features/vectorize.py
@@ -1,9 +1,10 @@
 import pandas as pd
 import pickle
+from pathlib import Path
 from sklearn.feature_extraction.text import TfidfVectorizer
 from sklearn.model_selection import train_test_split
 from src.data.preprocess_data import preprocess_data
-from config.config import DATA_PATH, VECTORIZER_PATH
+from config.config import BASE_DIR, DATA_PATH, VECTORIZER_PATH
 
 
 def split_dataset(df):
@@ -13,10 +14,17 @@ def split_dataset(df):
     return X_train.tolist(), X_test.tolist(), y_train, y_test
 
 
-def fit_vectorizer(X_train):
-    vectorizer = TfidfVectorizer().fit(X_train)
-    with open(VECTORIZER_PATH, "wb") as f:
-        pickle.dump(vectorizer, f)
+def fit_vectorizer(X_train=None):
+    if not VECTORIZER_PATH.exists():
+        temp = Path(BASE_DIR).joinpath("temp")
+        temp.mkdir(exist_ok=True)
+        vectorizer = TfidfVectorizer().fit(X_train)
+        with open(VECTORIZER_PATH, "wb") as f:
+            pickle.dump(vectorizer, f)
+
+    with open(VECTORIZER_PATH, "rb") as f:
+        vect = pickle.load(f)
+    return vect
 
 
 def transform_vectorizer(vectorizer, data):
@@ -28,9 +36,9 @@ if __name__ == "__main__":
     df = pd.read_csv(DATA_PATH, header=None)
     df = preprocess_data(df)
     X_train, X_test, y_train, y_test = split_dataset(df)
-    fit_vectorizer(X_train)
-    with open(VECTORIZER_PATH, "rb") as f:
-        vect = pickle.load(f)
+    vect = fit_vectorizer(X_train)
+    # with open(VECTORIZER_PATH, "rb") as f:
+    #     vect = pickle.load(f)
 
     X_train_vec = transform_vectorizer(vectorizer=vect, data=X_train)
     X_test_vec = transform_vectorizer(vectorizer=vect, data=X_test)
